+++ code/test/sixsq/nuvla/server/resources/nuvlabox_1_lifecycle_test.clj
@@ -694,7 +694,11 @@
                                               :swarm-endpoint      "https://swarm.example.com"
                                               :minio-access-key    "access"
                                               :minio-secret-key    "secret"
-                                              :minio-endpoint      "https://minio.example.com"}))
+                                              :minio-endpoint      "https://minio.example.com"
+                                              :kubernetes-client-key    "key"
+                                              :kubernetes-client-cert   "cert"
+                                              :kubernetes-client-ca     "ca"
+                                              :kubernetes-endpoint      "https://k8s.example.com"}))
               (ltu/body->edn)
               (ltu/is-status 200))
 
@@ -719,10 +723,10 @@
                                                                        "parent='%s'" isg-id)}))
                              (ltu/body->edn)
                              (ltu/is-status 200)
-                             (ltu/is-count 2)
+                             (ltu/is-count 3)
                              (ltu/entries))]
 
-            (is (= #{"swarm" "s3"} (set (map :subtype services))))
+            (is (= #{"swarm" "s3" "kubernetes"} (set (map :subtype services))))
 
             (doseq [{:keys [acl]} services]
               (is (= [nuvlabox-owner] (:view-acl acl))))
@@ -744,6 +748,8 @@
                 (if (= "s3" subtype)
                   (is (= 1 (count creds))))                 ;; only key/secret pair
 
+                (if (= "kubernetes" subtype)
+                  (is (= 1 (count creds))))
                 )))
 
           (-> session-beta
@@ -784,10 +790,10 @@
                                                                        "parent='%s'" isg-id)}))
                              (ltu/body->edn)
                              (ltu/is-status 200)
-                             (ltu/is-count 2)
+                             (ltu/is-count 3)
                              (ltu/entries))]
 
-            (is (= #{"swarm" "s3"} (set (map :subtype services))))
+            (is (= #{"swarm" "s3" "kubernetes"} (set (map :subtype services))))
 
             (doseq [{:keys [acl]} services]
               (is (= [nuvlabox-owner user-beta] (:view-acl acl))))
@@ -809,6 +815,8 @@
                 (if (= "s3" subtype)
                   (is (= 1 (count creds))))                 ;; only key/secret pair
 
+                (if (= "kubernetes" subtype)
+                  (is (= 1 (count creds))))
                 )))
 
           )))))